# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MicroVern API
  version: 2026-09-mvp
  description: Deterministic, pre-signing analysis for unsigned Algorand transaction groups.
servers:
  - url: http://localhost:4021
paths:
  /healthz:
    get:
      summary: Liveness probe
      responses:
        '200':
          description: Process is running.
  /readyz:
    get:
      summary: Payment readiness probe
      responses:
        '200':
          description: The configured facilitator supports the Testnet exact scheme.
        '503':
          $ref: '#/components/responses/Error'
  /v1/capabilities:
    get:
      summary: Discover supported inputs and payment configuration
      description: Includes the built-in versioned policy profiles available to callers.
      responses:
        '200':
          description: Service capabilities.
  /v1/validate-transaction:
    post:
      summary: Free bounded structural preflight
      description: Validates request shape and base64 bounds. It does not return an analysis report.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionRequest'
      responses:
        '200':
          description: Request shape is valid.
          content:
            application/json:
              schema:
                type: object
                required: [valid]
                properties:
                  valid:
                    const: true
        '400':
          $ref: '#/components/responses/Error'
        '413':
          $ref: '#/components/responses/Error'
  /v1/inspect-transaction:
    post:
      summary: Analyze an unsigned Algorand transaction group
      description: Requires x402 v2 exact USDC payment on the configured Algorand network. Send an Idempotency-Key to replay a completed report without another payment attempt for 10 minutes. A key is bound to the canonical unsigned transaction request and policy; reuse with different data returns 409 before payment handling.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            pattern: '^[A-Za-z0-9_-]{8,128}$'
          description: Generate one key per logical inspection request and reuse it only for a retry of that request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionRequest'
      responses:
        '200':
          description: Analysis report, or an idempotent replay of a completed paid report.
          headers:
            X-Request-Id:
              schema: { type: string }
            X-Idempotent-Replay:
              schema: { type: string, enum: ['true'] }
            X-MicroVern-Report-Id:
              schema: { type: string, pattern: '^[a-f0-9]{64}$' }
              description: Stable report ID equal to the report requestHash. It is not a retrieval token and cannot be used to access a report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InspectionReport'
        '400':
          $ref: '#/components/responses/Error'
        '402':
          description: x402 payment required. Requirements are encoded in the PAYMENT-REQUIRED header.
        '413':
          $ref: '#/components/responses/Error'
        '429':
          description: Too many unpaid inspection requests.
          headers:
            Retry-After:
              schema: { type: integer }
        '409':
          description: An inspection with this key is in progress, or the key is bound to a different request.
components:
  responses:
    Error:
      description: Client-visible validation or readiness error.
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error: { type: string }
  schemas:
    InspectionRequest:
      type: object
      required: [network, unsignedTransactionGroup]
      properties:
        network:
          type: string
          enum: [algorand-testnet, algorand-mainnet]
        unsignedTransactionGroup:
          type: string
          description: Base64 of one to sixteen concatenated algosdk.encodeUnsignedTransaction values. Decoded payload is limited to 65536 bytes.
        policy:
          type: object
          properties:
            maxAlgoSend: { type: number, minimum: 0 }
            maxUsdcSend: { type: number, minimum: 0 }
            allowRekey: { type: boolean }
            allowCloseOut: { type: boolean }
            allowUnknownApps: { type: boolean }
            allowedApplicationIds:
              type: array
              items: { type: integer, minimum: 0 }
            allowedAssetIds:
              type: array
              description: Explicit ASA allowlist. An empty array permits no ASA actions.
              items: { type: integer, minimum: 0 }
            prohibitAdminActions: { type: boolean }
        policyProfile:
          type: string
          enum: [strict-usdc-v1, algo-only-v1, no-admin-actions-v1]
          description: Built-in, versioned profile. Provide this instead of an ad-hoc policy.
        accountStateChecks:
          type: object
          additionalProperties: false
          required: [consent]
          description: Explicit consent for MicroVern to make read-only public Algod observations for sender accounts and ASAs touched by the group. Omit to make no account-state query.
          properties:
            consent: { const: true }
    InspectionReport:
      type: object
      required: [verdict, riskScore, summary, reviewSummary, actions, findings, policyEvaluation, rulesetVersion, disclaimer, requestHash, reportChecksum]
      properties:
        verdict: { type: string, enum: [allow, review, block] }
        riskScore: { type: integer, minimum: 0, maximum: 100 }
        summary: { type: string }
        reviewSummary:
          type: object
          required: [transactionCount, totalAlgoSent, totalUsdcSent, totalFeeAlgo, recipients, assetIds]
          properties:
            transactionCount: { type: integer, minimum: 1 }
            totalAlgoSent: { type: string, description: Excludes any ALGO close-out remainder, which is separately flagged. }
            totalUsdcSent: { type: string }
            totalFeeAlgo: { type: string }
            recipients:
              type: array
              items: { type: string }
            assetIds:
              type: array
              description: Algorand ASA IDs touched by asset-transfer actions, including opt-ins and opt-outs.
              items: { type: integer, minimum: 0 }
        actions:
          type: array
          items:
            type: object
            required: [index, type, description, consequences]
            properties:
              index: { type: integer, minimum: 0 }
              type: { type: string }
              description: { type: string }
              consequences:
                type: array
                items: { type: string }
              application:
                type: object
                description: Present for application calls. Exact registry matches are recognized; all others remain explicitly unknown.
                required: [registryVersion, applicationId, recognition]
                properties:
                  registryVersion: { type: string }
                  applicationId: { type: integer, minimum: 0 }
                  recognition: { type: string, enum: [recognized, known-application-unknown-method, unknown] }
                  name: { type: string }
                  method: { type: string }
                  referenceUrl: { type: string, format: uri }
        findings:
          type: array
          items:
            type: object
            required: [code, severity, transactionIndex, message]
            properties:
              code: { type: string }
              severity: { type: string, enum: [low, medium, high, critical] }
              transactionIndex: { type: integer, minimum: 0 }
              message: { type: string }
        policyEvaluation:
          type: object
          additionalProperties:
            type: string
            enum: [passed, failed, not-configured]
        policyProfile:
          type: object
          description: Present when the request selected a built-in versioned policy profile.
          required: [id, version]
          properties:
            id: { type: string }
            version: { type: string }
        accountState:
          type: object
          description: Present only after accountStateChecks consent. `observed` is a time-bounded public-node observation, not a guarantee that state will remain unchanged.
          required: [status, source, accounts, notice]
          properties:
            status: { type: string, enum: [observed, not-configured, unavailable] }
            source: { const: algod }
            observedRound: { type: integer, minimum: 0 }
            observedAt: { type: string, format: date-time }
            notice: { type: string }
            accounts:
              type: array
              items:
                type: object
                required: [address, active, assetOptIns]
                properties:
                  address: { type: string }
                  active: { type: boolean, description: True only when the configured Algod node returned an account ledger record. }
                  balanceMicroAlgos: { type: string, pattern: '^\d+$' }
                  assetOptIns:
                    type: array
                    items:
                      type: object
                      required: [assetId, optedIn]
                      properties:
                        assetId: { type: integer, minimum: 0 }
                        optedIn: { type: boolean }
        rulesetVersion: { type: string }
        disclaimer: { type: string }
        requestHash:
          type: string
          pattern: '^[a-f0-9]{64}$'
          description: Lowercase SHA-256 of the canonical unsigned transaction request and policy.
        reportChecksum:
          type: string
          pattern: '^[a-f0-9]{64}$'
          description: Lowercase SHA-256 binding the report content to requestHash.